chore: refactored util function into utils
